PUB-1691 Break up subscription service class (#152)
* Break up subscription service class

* Remove interface class for subscription service

* Fixed sonarqube complaints
