+++ src/main/java/uk/gov/hmcts/reform/pip/subscription/management/service/DataManagementService.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.reform.pip.subscription.management.service;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.reactive.function.client.WebClientException;
 import uk.gov.hmcts.reform.pip.subscription.management.models.subscriptionmanagement.Location;
@@ -12,7 +12,7 @@ import java.net.URI;
 import java.net.URISyntaxException;
 
 @Slf4j
-@Component
+@Service
 public class DataManagementService {
 
     @Autowired
