+++ src/main/java/uk/gov/hmcts/reform/pip/subscription/management/service/SubscriptionService.java
@@ -1,75 +1,142 @@
 package uk.gov.hmcts.reform.pip.subscription.management.service;
 
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import uk.gov.hmcts.reform.pip.model.enums.UserActions;
+import uk.gov.hmcts.reform.pip.subscription.management.errorhandling.exceptions.SubscriptionNotFoundException;
 import uk.gov.hmcts.reform.pip.subscription.management.models.Subscription;
-import uk.gov.hmcts.reform.pip.subscription.management.models.external.data.management.Artefact;
-import uk.gov.hmcts.reform.pip.subscription.management.models.response.UserSubscription;
+import uk.gov.hmcts.reform.pip.subscription.management.repository.SubscriptionRepository;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
-@SuppressWarnings({"PMD.TooManyMethods"})
-public interface SubscriptionService {
+import static uk.gov.hmcts.reform.pip.model.LogBuilder.writeLog;
+import static uk.gov.hmcts.reform.pip.subscription.management.models.SearchType.LOCATION_ID;
+
+/**
+ * Service layer for dealing with subscriptions.
+ */
+@Slf4j
+@Service
+public class SubscriptionService {
+
+    @Autowired
+    SubscriptionRepository repository;
+
+    @Autowired
+    DataManagementService dataManagementService;
+
+    public Subscription createSubscription(Subscription subscription, String actioningUserId) {
+        log.info(writeLog(actioningUserId, UserActions.CREATE_SUBSCRIPTION,
+                          subscription.getSearchType().toString()));
+
+        duplicateSubscriptionHandler(subscription);
+
+        subscription.setLastUpdatedDate(subscription.getCreatedDate());
+
+        if (subscription.getSearchType().equals(LOCATION_ID)) {
+            subscription.setLocationName(dataManagementService.getCourtName(subscription.getSearchValue()));
+        }
+        return repository.save(subscription);
+    }
+
+    public void configureListTypesForSubscription(String userId, List<String> listType) {
+        log.info(writeLog(userId, UserActions.CREATE_SUBSCRIPTION, LOCATION_ID.name()));
+
+        repository.updateLocationSubscriptions(userId,
+            listType == null ? "" : StringUtils.join(listType, ','));
+    }
+
+    public void deleteById(UUID id, String actioningUserId) {
+        Optional<Subscription> subscription = repository.findById(id);
+
+        if (subscription.isEmpty()) {
+            throw new SubscriptionNotFoundException(String.format(
+                "No subscription found with the subscription id %s",
+                id
+            ));
+        }
+
+        log.info(writeLog(actioningUserId, UserActions.DELETE_SUBSCRIPTION,
+                          id.toString()));
+
+        repository.deleteById(id);
+    }
+
+    public void bulkDeleteSubscriptions(List<UUID> ids) {
+        List<Subscription> subscriptions = repository.findByIdIn(ids);
+        if (ids.size() > subscriptions.size()) {
+            List<UUID> missingIds = new ArrayList<>(ids);
+            missingIds.removeAll(subscriptions.stream()
+                                     .map(Subscription::getId)
+                                     .toList());
+            throw new SubscriptionNotFoundException("No subscription found with the subscription ID(s): "
+                    + missingIds.toString().replace("[", "").replace("]", ""));
+        }
+
+        repository.deleteByIdIn(ids);
+        subscriptions.forEach(s -> log.info(writeLog(s.getUserId(), UserActions.DELETE_SUBSCRIPTION,
+                                                     s.getId().toString())));
+    }
+
+    public List<Subscription> findAll() {
+        return repository.findAll();
+    }
+
+    public Subscription findById(UUID subscriptionId) {
+        Optional<Subscription> subscription = repository.findById(subscriptionId);
+        if (subscription.isEmpty()) {
+            throw new SubscriptionNotFoundException(String.format(
+                "No subscription found with the subscription id %s",
+                subscriptionId
+            ));
+        }
+        return subscription.get();
+    }
 
     /**
-     * Method to create a new subscription.
-     * @param subscription The subscription to create.
-     * @param actioningUserId The ID of the user creating the subscription.
-     * @return The created subscription.
+     * Take in a new user subscription and check if any with the same criteria already exist.
+     * If it does then delete the original subscription as the new one will supersede it.
+     *
+     * @param subscription The new subscription that will be created
      */
-    Subscription createSubscription(Subscription subscription, String actioningUserId);
+    private void duplicateSubscriptionHandler(Subscription subscription) {
+        repository.findByUserId(subscription.getUserId()).forEach(existingSub -> {
+            if (existingSub.getSearchType().equals(subscription.getSearchType())
+                && existingSub.getSearchValue().equals(subscription.getSearchValue())) {
+                repository.delete(existingSub);
+            }
+        });
+    }
 
-    /**
-     * Delete subscription by id.
-     * @param id The id of the subscription to delete.
-     * @param actioningUserId The id of the user deleting the subscription.
-     */
-    void deleteById(UUID id, String actioningUserId);
+    public String getAllSubscriptionsDataForMiReporting() {
+        StringBuilder builder = new StringBuilder(48);
+        builder.append("id,channel,search_type,user_id,court_name").append(System.lineSeparator());
+        repository.getAllSubsDataForMi()
+            .forEach(line -> builder.append(line).append(System.lineSeparator()));
+        return builder.toString();
+    }
 
-    /**
-     * Bulk delete subscriptions by IDs.
-     * @param ids the list of IDs of the subscriptions to delete.
-     */
-    void bulkDeleteSubscriptions(List<UUID> ids);
+    public String getLocalSubscriptionsDataForMiReporting() {
+        StringBuilder builder = new StringBuilder(49);
+        builder.append("id,search_value,channel,user_id,court_name").append(System.lineSeparator());
+        repository.getLocalSubsDataForMi()
+            .forEach(line -> builder.append(line).append(System.lineSeparator()));
+        return builder.toString();
+    }
 
-    /**
-     * Find all subscriptions.
-     * @return The list of subscriptions that have been found.
-     */
-    List<Subscription> findAll();
-
-    /**
-     * Find subscription by id.
-     * @param subscriptionId The id of the subscription to be found.
-     * @return The subscription that has been found.
-     */
-    Subscription findById(UUID subscriptionId);
-
-    /**
-     * Find all subscriptions for a given user.
-     * @return The list of subscriptions that have been found.
-     */
-    UserSubscription findByUserId(String userId);
-
-    void collectSubscribers(Artefact artefact);
-
-    void collectThirdPartyForDeletion(Artefact artefact);
-
-    String getAllSubscriptionsDataForMiReporting();
-
-    String getLocalSubscriptionsDataForMiReporting();
-
-    /**
-     * Method to update an existing subscription.
-     * @param userId and ListType to update.
-     */
-    void configureListTypesForSubscription(String userId, List<String> listType);
-
-    /**
-     * Delete all subscriptions by the user id.
-     * @param userId The user id to delete the subscriptions from.
-     * @return A confirmation message.
-     */
-    String deleteAllByUserId(String userId);
-
-    List<Subscription> findSubscriptionsByLocationId(String value);
+    public List<Subscription> findSubscriptionsByLocationId(String value) {
+        List<Subscription> locationSubscriptions = repository.findSubscriptionsByLocationId(value);
+        if (locationSubscriptions.isEmpty()) {
+            throw new SubscriptionNotFoundException(String.format(
+                "No subscription found with the location id %s",
+                value
+            ));
+        }
+        return repository.findSubscriptionsByLocationId(value);
+    }
 }
