+++ src/test/java/uk/gov/hmcts/reform/pip/subscription/management/service/SubscriptionServiceTest.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.reform.pip.subscription.management.service;
 
-import nl.altindag.log.LogCaptor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -14,70 +13,35 @@ import uk.gov.hmcts.reform.pip.subscription.management.errorhandling.exceptions.
 import uk.gov.hmcts.reform.pip.subscription.management.models.Channel;
 import uk.gov.hmcts.reform.pip.subscription.management.models.SearchType;
 import uk.gov.hmcts.reform.pip.subscription.management.models.Subscription;
-import uk.gov.hmcts.reform.pip.subscription.management.models.SubscriptionsSummary;
-import uk.gov.hmcts.reform.pip.subscription.management.models.SubscriptionsSummaryDetails;
-import uk.gov.hmcts.reform.pip.subscription.management.models.external.data.management.Artefact;
-import uk.gov.hmcts.reform.pip.subscription.management.models.external.data.management.ListType;
-import uk.gov.hmcts.reform.pip.subscription.management.models.external.data.management.Sensitivity;
-import uk.gov.hmcts.reform.pip.subscription.management.models.external.publication.services.ThirdPartySubscription;
-import uk.gov.hmcts.reform.pip.subscription.management.models.external.publication.services.ThirdPartySubscriptionArtefact;
-import uk.gov.hmcts.reform.pip.subscription.management.models.response.CaseSubscription;
-import uk.gov.hmcts.reform.pip.subscription.management.models.response.ListTypeSubscription;
-import uk.gov.hmcts.reform.pip.subscription.management.models.response.LocationSubscription;
-import uk.gov.hmcts.reform.pip.subscription.management.models.response.UserSubscription;
 import uk.gov.hmcts.reform.pip.subscription.management.repository.SubscriptionRepository;
 
-import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.concurrent.ConcurrentHashMap;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.lenient;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.pip.subscription.management.helpers.SubscriptionUtils.createMockSubscription;
 import static uk.gov.hmcts.reform.pip.subscription.management.helpers.SubscriptionUtils.createMockSubscriptionList;
 import static uk.gov.hmcts.reform.pip.subscription.management.helpers.SubscriptionUtils.findableSubscription;
+import static uk.gov.hmcts.reform.pip.subscription.management.models.external.data.management.ListType.CIVIL_AND_FAMILY_DAILY_CAUSE_LIST;
+import static uk.gov.hmcts.reform.pip.subscription.management.models.external.data.management.ListType.CIVIL_DAILY_CAUSE_LIST;
 
 @ActiveProfiles("non-async")
 @ExtendWith({MockitoExtension.class})
-@SuppressWarnings({"PMD.LawOfDemeter", "PMD.TooManyFields", "PMD.ExcessiveImports"})
 class SubscriptionServiceTest {
     private static final String USER_ID = "Ralph21";
-    private static final String USER_ID_NO_SUBS = "Tina21";
     private static final String SEARCH_VALUE = "193254";
-    private static final String CASE_ID = "123";
-    private static final String URN = "312";
-    private static final String CASE_NAME = "case-name";
     private static final Channel EMAIL = Channel.EMAIL;
-    private static final String COURT_MATCH = "1";
-    private static final String CASE_MATCH = "case match";
-    private static final String ACCEPTED_USER_ID = "2";
-    private static final String FORBIDDEN_USER_ID = "3";
-    private static final String SUBSCRIBER_NOTIFICATION_LOG = "Summary being sent to publication services for id";
-    private static final String LOG_MESSAGE_MATCH = "Log messages should match.";
-    private static final String CASE_NUMBER_KEY = "caseNumber";
-    private static final String CASE_URN_KEY = "caseUrn";
-    private static final UUID TEST_UUID = UUID.randomUUID();
-    private static final String TEST_USER_EMAIL = "a@b.com";
-    private static final String SUCCESS = "Success";
-    private static final String TEST = "test";
-    private static final String LOCATION_ID = "1";
     private static final String ACTIONING_USER_ID = "1234-1234";
 
     public static final List<String> EXAMPLE_CSV_ALL = List.of(
@@ -91,24 +55,14 @@ class SubscriptionServiceTest {
         "34edfcde-4546-46b8-98e6-2717da3185e8,3,API,2fe899ff-96ed-435a-bcad-1411bbe96d2a,Oxford Combined Court Centre");
 
     private static final String COURT_NAME = "test court name";
+    private static final LocalDateTime DATE_ADDED = LocalDateTime.now();
 
     private static final String SUBSCRIPTION_CREATED_ERROR = "The returned subscription does "
         + "not match the expected subscription";
 
     private List<Subscription> mockSubscriptionList;
     private Subscription mockSubscription;
-    private final SubscriptionsSummary mockSubscriptionsSummary = new SubscriptionsSummary();
-    private final SubscriptionsSummaryDetails mockSubscriptionsSummaryDetails = new SubscriptionsSummaryDetails();
     private Subscription findableSubscription;
-    private LocalDateTime dateAdded;
-
-    private final Artefact classifiedArtefactMatches = new Artefact();
-    private final Artefact publicArtefactMatches = new Artefact();
-    private final Subscription returnedSubscription = new Subscription();
-    private final Subscription restrictedSubscription = new Subscription();
-    private final List<Object> cases = new ArrayList<>();
-    private final Map<String, List<Object>> searchTerms = new ConcurrentHashMap<>();
-    private final Map<String, List<Subscription>> returnedMappedEmails = new ConcurrentHashMap<>();
 
     @Captor
     private ArgumentCaptor<ArrayList<UUID>> listCaptor;
@@ -117,60 +71,18 @@ class SubscriptionServiceTest {
     DataManagementService dataManagementService;
 
     @Mock
-    PublicationServicesService publicationServicesService;
-
-    @Mock
     SubscriptionRepository subscriptionRepository;
 
-    @Mock
-    AccountManagementService accountManagementService;
-
-    @Mock
-    ChannelManagementService channelManagementService;
-
     @InjectMocks
-    SubscriptionServiceImpl subscriptionService;
+    SubscriptionService subscriptionService;
 
     @BeforeEach
     void setup() {
-        LinkedHashMap<String, String> map = new LinkedHashMap<>();
-        LinkedHashMap<String, String> map2 = new LinkedHashMap<>();
-        map.put(CASE_NUMBER_KEY, CASE_MATCH);
-        map.put(CASE_URN_KEY, TEST);
-        map2.put(CASE_NUMBER_KEY, TEST);
-        map2.put(CASE_URN_KEY, CASE_MATCH);
-
-        cases.add(map);
-        cases.add(map2);
-
-        searchTerms.put("cases", cases);
-        classifiedArtefactMatches.setArtefactId(TEST_UUID);
-        classifiedArtefactMatches.setSensitivity(Sensitivity.CLASSIFIED);
-        classifiedArtefactMatches.setSearch(searchTerms);
-        classifiedArtefactMatches.setLocationId(COURT_MATCH);
-        classifiedArtefactMatches.setListType(ListType.SJP_PRESS_LIST);
-
-        publicArtefactMatches.setArtefactId(TEST_UUID);
-        publicArtefactMatches.setSensitivity(Sensitivity.PUBLIC);
-        publicArtefactMatches.setLocationId(COURT_MATCH);
-        publicArtefactMatches.setSearch(searchTerms);
-        publicArtefactMatches.setListType(ListType.MAGISTRATES_PUBLIC_LIST);
-
-        returnedSubscription.setUserId(ACCEPTED_USER_ID);
-        restrictedSubscription.setUserId(FORBIDDEN_USER_ID);
-
-        dateAdded = LocalDateTime.now();
-        mockSubscription = createMockSubscription(USER_ID, SEARCH_VALUE, EMAIL, dateAdded,
-                                                  ListType.CIVIL_DAILY_CAUSE_LIST);
-        mockSubscriptionList = createMockSubscriptionList(dateAdded);
+        mockSubscription = createMockSubscription(USER_ID, SEARCH_VALUE, EMAIL, DATE_ADDED,
+                                                  CIVIL_DAILY_CAUSE_LIST);
+        mockSubscriptionList = createMockSubscriptionList(DATE_ADDED);
         findableSubscription = findableSubscription();
-
-        mockSubscriptionsSummary.setEmail(TEST_USER_EMAIL);
-        mockSubscriptionsSummary.setArtefactId(TEST_UUID);
         mockSubscription.setChannel(Channel.EMAIL);
-
-        lenient().when(subscriptionRepository.findByUserId(USER_ID)).thenReturn(mockSubscriptionList);
-        lenient().when(subscriptionRepository.findByUserId(USER_ID_NO_SUBS)).thenReturn(new ArrayList<>());
     }
 
     @Test
@@ -210,8 +122,7 @@ class SubscriptionServiceTest {
         when(dataManagementService.getCourtName(SEARCH_VALUE)).thenReturn(COURT_NAME);
         when(subscriptionRepository.save(mockSubscription)).thenReturn(mockSubscription);
         assertEquals(subscriptionService.createSubscription(mockSubscription, ACTIONING_USER_ID), mockSubscription,
-                     SUBSCRIPTION_CREATED_ERROR
-        );
+                     SUBSCRIPTION_CREATED_ERROR);
     }
 
     @Test
@@ -221,20 +132,17 @@ class SubscriptionServiceTest {
         when(dataManagementService.getCourtName(SEARCH_VALUE)).thenReturn(COURT_NAME);
         when(subscriptionRepository.save(mockSubscription)).thenReturn(mockSubscription);
         assertEquals(subscriptionService.createSubscription(mockSubscription, ACTIONING_USER_ID), mockSubscription,
-                     SUBSCRIPTION_CREATED_ERROR
-        );
+                     SUBSCRIPTION_CREATED_ERROR);
     }
 
     @Test
     void testCreateSubscriptionWithCourtNameWithMultipleListType() {
         mockSubscription.setSearchType(SearchType.LOCATION_ID);
-        mockSubscription.setListType(List.of(ListType.CIVIL_DAILY_CAUSE_LIST.toString(),
-                                             ListType.CIVIL_AND_FAMILY_DAILY_CAUSE_LIST.toString()));
+        mockSubscription.setListType(List.of(CIVIL_DAILY_CAUSE_LIST.name(), CIVIL_AND_FAMILY_DAILY_CAUSE_LIST.name()));
         when(dataManagementService.getCourtName(SEARCH_VALUE)).thenReturn(COURT_NAME);
         when(subscriptionRepository.save(mockSubscription)).thenReturn(mockSubscription);
         assertEquals(subscriptionService.createSubscription(mockSubscription, ACTIONING_USER_ID), mockSubscription,
-                     SUBSCRIPTION_CREATED_ERROR
-        );
+                     SUBSCRIPTION_CREATED_ERROR);
     }
 
     @Test
@@ -256,7 +164,7 @@ class SubscriptionServiceTest {
     @Test
     void testConfigureListTypesForLocationSubscription() {
         doNothing().when(subscriptionRepository).updateLocationSubscriptions(any(), any());
-        subscriptionService.configureListTypesForSubscription(USER_ID, List.of(ListType.CIVIL_DAILY_CAUSE_LIST.name()));
+        subscriptionService.configureListTypesForSubscription(USER_ID, List.of(CIVIL_DAILY_CAUSE_LIST.name()));
 
         assertEquals(USER_ID, mockSubscription.getUserId(),
                      SUBSCRIPTION_CREATED_ERROR);
@@ -346,373 +254,6 @@ class SubscriptionServiceTest {
     }
 
     @Test
-    void testNoSubscriptionsReturnsEmpty() {
-        assertEquals(new UserSubscription(), subscriptionService.findByUserId(USER_ID_NO_SUBS),
-                     "Should return empty user subscriptions");
-    }
-
-    @Test
-    void testFindByUserIdOnlyCourt() {
-        mockSubscription.setSearchType(SearchType.LOCATION_ID);
-        mockSubscription.setLocationName("Test court");
-        when(subscriptionRepository.findByUserId(USER_ID)).thenReturn(List.of(mockSubscription));
-        LocationSubscription expected = new LocationSubscription();
-        expected.setSubscriptionId(mockSubscription.getId());
-        expected.setLocationName("Test court");
-        expected.setLocationId("193254");
-        expected.setListType(List.of(ListType.CIVIL_DAILY_CAUSE_LIST.name()));
-        expected.setDateAdded(dateAdded);
-
-        UserSubscription result = subscriptionService.findByUserId(USER_ID);
-
-        assertEquals(List.of(expected), result.getLocationSubscriptions(),
-                     "Should return court name");
-        assertEquals(0, result.getCaseSubscriptions().size(), "Cases should be empty");
-    }
-
-    @Test
-    void testFindByUserIdCaseType() {
-        mockSubscription.setSearchType(SearchType.CASE_ID);
-        mockSubscription.setCaseNumber(CASE_ID);
-        mockSubscription.setCaseName(CASE_NAME);
-        mockSubscription.setUrn(URN);
-
-        CaseSubscription expected = new CaseSubscription();
-        expected.setCaseNumber(CASE_ID);
-        expected.setCaseName(CASE_NAME);
-        expected.setUrn(URN);
-        expected.setSubscriptionId(mockSubscription.getId());
-        expected.setDateAdded(dateAdded);
-        when(subscriptionRepository.findByUserId(USER_ID)).thenReturn(List.of(mockSubscription));
-
-        assertEquals(List.of(expected), subscriptionService.findByUserId(USER_ID).getCaseSubscriptions(),
-                     "Should return populated case");
-    }
-
-    @Test
-    void testFindByUserIdLength() {
-        UserSubscription result = subscriptionService.findByUserId(USER_ID);
-        assertEquals(6, result.getCaseSubscriptions().size(),
-                     "Should add all CaseSubscriptions to UserSubscriptions");
-        assertEquals(2, result.getLocationSubscriptions().size(), "Should add all court names");
-    }
-
-    @Test
-    void testFindByUserId() {
-        UserSubscription result = subscriptionService.findByUserId(USER_ID);
-        for (int i = 0; i < 6; i++) {
-            assertEquals(CASE_ID + i, result.getCaseSubscriptions().get(i).getCaseNumber(),
-                         "Should contain correct caseNumber");
-        }
-        assertEquals(COURT_NAME, result.getLocationSubscriptions().get(0).getLocationName(),
-                     "Should match court name");
-    }
-
-    @Test
-    void testFindByUserIdCreatedDates() {
-        UserSubscription result = subscriptionService.findByUserId(USER_ID);
-        for (int i = 0; i < 6; i++) {
-            assertEquals(dateAdded, result.getCaseSubscriptions().get(i).getDateAdded(),
-                         "Should match dateAdded");
-        }
-        assertEquals(dateAdded, result.getLocationSubscriptions().get(0).getDateAdded(), "Should match dateAdded");
-    }
-
-    @Test
-    void testFindByUserIdAssignsIdForCourt() {
-        mockSubscription.setSearchType(SearchType.LOCATION_ID);
-        when(subscriptionRepository.findByUserId(USER_ID)).thenReturn(List.of(mockSubscription));
-
-        assertEquals(mockSubscription.getId(),
-                     subscriptionService.findByUserId(USER_ID).getLocationSubscriptions().get(0).getSubscriptionId(),
-                     "Should match subscriptionId");
-    }
-
-    @Test
-    void testFindByUserIdAssignsIdForCase() {
-        mockSubscription.setSearchType(SearchType.CASE_ID);
-        when(subscriptionRepository.findByUserId(USER_ID)).thenReturn(List.of(mockSubscription));
-
-        assertEquals(mockSubscription.getId(),
-                     subscriptionService.findByUserId(USER_ID).getCaseSubscriptions().get(0).getSubscriptionId(),
-                     "Should match subscriptionId");
-    }
-
-    @Test
-    void testFindByUserIdListTypeSubscriptions() {
-        mockSubscription.setSearchType(SearchType.LIST_TYPE);
-        mockSubscription.setSearchValue(ListType.CARE_STANDARDS_LIST.toString());
-
-        when(subscriptionRepository.findByUserId(USER_ID)).thenReturn(List.of(mockSubscription));
-
-        List<ListTypeSubscription> listTypeSubscriptions =
-            subscriptionService.findByUserId(USER_ID).getListTypeSubscriptions();
-
-
-        assertEquals(1, listTypeSubscriptions.size(), "Unexpected number of list type subscriptions returned");
-
-        assertEquals(mockSubscription.getId(), listTypeSubscriptions.get(0).getSubscriptionId(),
-                     "Returned list type subscription does not match ID");
-        assertEquals(mockSubscription.getSearchValue(), listTypeSubscriptions.get(0).getListType(),
-                     "Returned list type subscription does not match list type");
-        assertEquals(mockSubscription.getChannel(), listTypeSubscriptions.get(0).getChannel(),
-                     "Returned list type subscription does not match channel");
-        assertEquals(mockSubscription.getCreatedDate(), listTypeSubscriptions.get(0).getDateAdded(),
-                     "Returned list type subscription does not match created date");
-    }
-
-    @Test
-    void testFindByUserIdNoListTypeSubscriptions() {
-        mockSubscription.setSearchType(SearchType.CASE_ID);
-        when(subscriptionRepository.findByUserId(USER_ID)).thenReturn(List.of(mockSubscription));
-
-        assertEquals(0, subscriptionService.findByUserId(USER_ID).getListTypeSubscriptions().size(),
-                     "Unexpected number of list type subscriptions returned");
-    }
-
-    @Test
-    void testCollectSubscribersCourtSubscriptionNotClassified() throws IOException {
-        returnedSubscription.setChannel(Channel.EMAIL);
-        returnedMappedEmails.put(TEST_USER_EMAIL, List.of(returnedSubscription));
-        when(subscriptionRepository.findSubscriptionsByLocationSearchValue(SearchType.LOCATION_ID.toString(),
-            COURT_MATCH, ListType.MAGISTRATES_PUBLIC_LIST.name()))
-                .thenReturn(List.of(returnedSubscription));
-        when(channelManagementService.getMappedEmails(any())).thenReturn(returnedMappedEmails);
-        when(publicationServicesService.postSubscriptionSummaries(any(), any(), any())).thenReturn(SUCCESS);
-        try (LogCaptor logCaptor = LogCaptor.forClass(SubscriptionServiceImpl.class)) {
-            subscriptionService.collectSubscribers(publicArtefactMatches);
-            assertTrue(logCaptor.getInfoLogs().get(0).contains(SUBSCRIBER_NOTIFICATION_LOG),
-                       LOG_MESSAGE_MATCH);
-        } catch (Exception ex) {
-            throw new IOException(ex.getMessage());
-        }
-    }
-
-    @Test
-    void testCollectSubscribersCourtSubscriptionWithListTypeNotClassified() throws IOException {
-        returnedSubscription.setChannel(Channel.EMAIL);
-        returnedSubscription.setListType(List.of(ListType.MAGISTRATES_PUBLIC_LIST.toString(),
-                                             ListType.CIVIL_DAILY_CAUSE_LIST.toString()));
-        returnedMappedEmails.put(TEST_USER_EMAIL, List.of(returnedSubscription));
-        when(subscriptionRepository.findSubscriptionsByLocationSearchValue(SearchType.LOCATION_ID.toString(),
-            COURT_MATCH, ListType.MAGISTRATES_PUBLIC_LIST.name()))
-                .thenReturn(List.of(returnedSubscription));
-        when(channelManagementService.getMappedEmails(any())).thenReturn(returnedMappedEmails);
-        when(publicationServicesService.postSubscriptionSummaries(any(), any(), any())).thenReturn(SUCCESS);
-        try (LogCaptor logCaptor = LogCaptor.forClass(SubscriptionServiceImpl.class)) {
-            subscriptionService.collectSubscribers(publicArtefactMatches);
-            assertTrue(logCaptor.getInfoLogs().get(0).contains(SUBSCRIBER_NOTIFICATION_LOG),
-                       LOG_MESSAGE_MATCH);
-        } catch (Exception ex) {
-            throw new IOException(ex.getMessage());
-        }
-    }
-
-    @Test
-    void testCollectSubscribersCaseUrn() {
-        mockSubscription.setSearchType(SearchType.CASE_URN);
-        mockSubscription.setSearchValue(CASE_URN_KEY);
-
-        mockSubscriptionsSummaryDetails.addToCaseUrn(CASE_URN_KEY);
-        mockSubscriptionsSummary.setSubscriptions(mockSubscriptionsSummaryDetails);
-
-        when(subscriptionRepository.findSubscriptionsByLocationSearchValue(SearchType.LOCATION_ID.toString(),
-                                                                   COURT_MATCH,
-                                                                   ListType.MAGISTRATES_PUBLIC_LIST.name()))
-            .thenReturn(List.of(mockSubscription));
-
-        returnedMappedEmails.put(TEST_USER_EMAIL, List.of(mockSubscription));
-
-        when(channelManagementService.getMappedEmails(List.of(mockSubscription))).thenReturn(returnedMappedEmails);
-
-        when(publicationServicesService.postSubscriptionSummaries(publicArtefactMatches.getArtefactId(),
-                                                                               TEST_USER_EMAIL,
-                                                                  List.of(mockSubscription))).thenReturn(SUCCESS);
-
-        try (LogCaptor logCaptor = LogCaptor.forClass(SubscriptionServiceImpl.class)) {
-            subscriptionService.collectSubscribers(publicArtefactMatches);
-
-            assertTrue(logCaptor.getInfoLogs().get(0).contains(SUBSCRIBER_NOTIFICATION_LOG),
-                       LOG_MESSAGE_MATCH);
-        }
-    }
-
-    @Test
-    void testCollectSubscribersCaseId() {
-        mockSubscription.setSearchType(SearchType.CASE_ID);
-        mockSubscription.setSearchValue(CASE_ID);
-
-        mockSubscriptionsSummaryDetails.addToCaseNumber(CASE_ID);
-        mockSubscriptionsSummary.setSubscriptions(mockSubscriptionsSummaryDetails);
-
-        when(subscriptionRepository.findSubscriptionsByLocationSearchValue(SearchType.LOCATION_ID.toString(),
-                                                                   COURT_MATCH,
-                                                                   ListType.MAGISTRATES_PUBLIC_LIST.name()))
-            .thenReturn(List.of(mockSubscription));
-
-        returnedMappedEmails.put(TEST_USER_EMAIL, List.of(mockSubscription));
-
-        when(channelManagementService.getMappedEmails(List.of(mockSubscription))).thenReturn(returnedMappedEmails);
-
-        when(publicationServicesService.postSubscriptionSummaries(publicArtefactMatches.getArtefactId(),
-                                                                               TEST_USER_EMAIL,
-                                                                  List.of(mockSubscription))).thenReturn(SUCCESS);
-
-        try (LogCaptor logCaptor = LogCaptor.forClass(SubscriptionServiceImpl.class)) {
-            subscriptionService.collectSubscribers(publicArtefactMatches);
-
-            assertTrue(logCaptor.getInfoLogs().get(0).contains(SUBSCRIBER_NOTIFICATION_LOG),
-                       LOG_MESSAGE_MATCH);
-        }
-    }
-
-    @Test
-    void testCollectSubscribersLocationId() {
-        mockSubscription.setSearchType(SearchType.LOCATION_ID);
-        mockSubscription.setSearchValue(COURT_MATCH);
-
-        mockSubscriptionsSummaryDetails.addToLocationId(COURT_MATCH);
-        mockSubscriptionsSummary.setSubscriptions(mockSubscriptionsSummaryDetails);
-
-        when(subscriptionRepository.findSubscriptionsByLocationSearchValue(SearchType.LOCATION_ID.toString(),
-                                                                   COURT_MATCH,
-                                                                   ListType.MAGISTRATES_PUBLIC_LIST.name()))
-            .thenReturn(List.of(mockSubscription));
-
-        returnedMappedEmails.put(TEST_USER_EMAIL, List.of(mockSubscription));
-
-        when(channelManagementService.getMappedEmails(List.of(mockSubscription))).thenReturn(returnedMappedEmails);
-
-        when(publicationServicesService.postSubscriptionSummaries(publicArtefactMatches.getArtefactId(),
-                                                                               TEST_USER_EMAIL,
-                                                                  List.of(mockSubscription))).thenReturn(SUCCESS);
-
-        try (LogCaptor logCaptor = LogCaptor.forClass(SubscriptionServiceImpl.class)) {
-            subscriptionService.collectSubscribers(publicArtefactMatches);
-
-            assertTrue(logCaptor.getInfoLogs().get(0).contains(SUBSCRIBER_NOTIFICATION_LOG),
-                       LOG_MESSAGE_MATCH);
-        }
-    }
-
-    @Test
-    void testCollectListTypeSubscription() throws IOException {
-        mockSubscription.setSearchType(SearchType.LIST_TYPE);
-        mockSubscription.setSearchValue(ListType.MAGISTRATES_PUBLIC_LIST.toString());
-        returnedMappedEmails.put(TEST_USER_EMAIL, List.of(mockSubscription));
-        lenient().when(subscriptionRepository.findSubscriptionsBySearchValue(SearchType.LIST_TYPE.toString(),
-                                                                   ListType.MAGISTRATES_PUBLIC_LIST.toString()))
-            .thenReturn(List.of(mockSubscription));
-        when(channelManagementService.getMappedEmails(List.of(mockSubscription))).thenReturn(returnedMappedEmails);
-        lenient().when(publicationServicesService.postSubscriptionSummaries(any(), any(), any())).thenReturn(SUCCESS);
-
-        try (LogCaptor logCaptor = LogCaptor.forClass(SubscriptionServiceImpl.class)) {
-            subscriptionService.collectSubscribers(publicArtefactMatches);
-            assertTrue(logCaptor.getInfoLogs().get(0).contains(SUBSCRIBER_NOTIFICATION_LOG),
-                       LOG_MESSAGE_MATCH);
-        } catch (Exception ex) {
-            throw new IOException(ex.getMessage());
-        }
-    }
-
-    @Test
-    void testCollectApiSubscribers() throws IOException {
-        mockSubscription.setChannel(Channel.API_COURTEL);
-        Map<String, List<Subscription>> returnedMap = new ConcurrentHashMap<>();
-        returnedMap.put(TEST, List.of(mockSubscription));
-        ThirdPartySubscription thirdPartySubscription = new ThirdPartySubscription(TEST, TEST_UUID);
-        when(subscriptionRepository.findSubscriptionsByLocationSearchValue(SearchType.LOCATION_ID.toString(),
-                                                                   COURT_MATCH,
-                                                                   ListType.MAGISTRATES_PUBLIC_LIST.name()))
-            .thenReturn(List.of(mockSubscription));
-        when(channelManagementService.getMappedApis(List.of(mockSubscription))).thenReturn(returnedMap);
-        when(publicationServicesService.sendThirdPartyList(thirdPartySubscription)).thenReturn(SUCCESS);
-        try (LogCaptor logCaptor = LogCaptor.forClass(SubscriptionServiceImpl.class)) {
-            subscriptionService.collectSubscribers(publicArtefactMatches);
-            assertTrue(logCaptor.getInfoLogs().get(0).contains(SUCCESS),
-                       LOG_MESSAGE_MATCH);
-        } catch (Exception ex) {
-            throw new IOException(ex.getMessage());
-        }
-    }
-
-    @Test
-    void testCollectSubscribersRestrictsClassified() {
-        returnedSubscription.setChannel(Channel.EMAIL);
-        returnedSubscription.setUserId(ACCEPTED_USER_ID);
-        returnedSubscription.setSearchType(SearchType.CASE_ID);
-        returnedSubscription.setSearchValue(CASE_ID);
-        restrictedSubscription.setChannel(Channel.EMAIL);
-        restrictedSubscription.setUserId(FORBIDDEN_USER_ID);
-        returnedSubscription.setSearchType(SearchType.CASE_ID);
-        returnedSubscription.setSearchValue(CASE_ID);
-
-        mockSubscriptionsSummary.setArtefactId(classifiedArtefactMatches.getArtefactId());
-        mockSubscriptionsSummaryDetails.addToCaseNumber(CASE_ID);
-        mockSubscriptionsSummary.setSubscriptions(mockSubscriptionsSummaryDetails);
-
-        lenient().when(subscriptionRepository.findSubscriptionsBySearchValue(SearchType.CASE_ID.name(),
-                                                                             CASE_MATCH))
-            .thenReturn(List.of(returnedSubscription, restrictedSubscription));
-
-        when(accountManagementService.isUserAuthorised(
-            ACCEPTED_USER_ID, ListType.SJP_PRESS_LIST, Sensitivity.CLASSIFIED)).thenReturn(true);
-        when(accountManagementService.isUserAuthorised(
-            FORBIDDEN_USER_ID, ListType.SJP_PRESS_LIST, Sensitivity.CLASSIFIED))
-            .thenReturn(false);
-
-        returnedMappedEmails.put(TEST_USER_EMAIL, List.of(returnedSubscription));
-        when(channelManagementService.getMappedEmails(List.of(returnedSubscription)))
-            .thenReturn(returnedMappedEmails);
-
-        when(publicationServicesService.postSubscriptionSummaries(publicArtefactMatches.getArtefactId(),
-                                                                  TEST_USER_EMAIL,
-                                                                  List.of(returnedSubscription))).thenReturn(SUCCESS);
-
-        try (LogCaptor logCaptor = LogCaptor.forClass(SubscriptionServiceImpl.class)) {
-            subscriptionService.collectSubscribers(classifiedArtefactMatches);
-
-            assertTrue(logCaptor.getInfoLogs().get(0).contains(SUBSCRIBER_NOTIFICATION_LOG), LOG_MESSAGE_MATCH);
-        }
-    }
-
-    @Test
-    void testCollectThirdPartyForDeletion() {
-        mockSubscription.setChannel(Channel.API_COURTEL);
-        Map<String, List<Subscription>> returnedMap = Collections.singletonMap(TEST, List.of(mockSubscription));
-        when(subscriptionRepository.findSubscriptionsBySearchValue(SearchType.LIST_TYPE.toString(),
-                                                                   publicArtefactMatches.getListType().name()))
-            .thenReturn(List.of(mockSubscription));
-        when(channelManagementService.getMappedApis(List.of(mockSubscription))).thenReturn(returnedMap);
-        ThirdPartySubscriptionArtefact subscriptionArtefact = new ThirdPartySubscriptionArtefact(
-            TEST, publicArtefactMatches);
-        when(publicationServicesService.sendEmptyArtefact(subscriptionArtefact))
-            .thenReturn(SUBSCRIBER_NOTIFICATION_LOG);
-        try (LogCaptor logCaptor = LogCaptor.forClass(SubscriptionServiceImpl.class)) {
-            subscriptionService.collectThirdPartyForDeletion(publicArtefactMatches);
-            assertTrue(logCaptor.getInfoLogs().get(0).contains(SUBSCRIBER_NOTIFICATION_LOG),
-                       LOG_MESSAGE_MATCH);
-        }
-    }
-
-    @Test
-    void testCollectThirdPartyForDeletionClassifiedExcluded() {
-        mockSubscription.setChannel(Channel.API_COURTEL);
-        Map<String, List<Subscription>> returnedMap = new ConcurrentHashMap<>();
-        returnedMap.put(TEST, List.of(mockSubscription));
-        lenient().when(subscriptionRepository.findSubscriptionsBySearchValue(SearchType.LIST_TYPE.toString(),
-                                                                   classifiedArtefactMatches.getListType().name()))
-            .thenReturn(List.of(mockSubscription));
-        when(accountManagementService.isUserAuthorised(mockSubscription.getUserId(),
-                                                       classifiedArtefactMatches.getListType(),
-                                                       classifiedArtefactMatches.getSensitivity())).thenReturn(false);
-        subscriptionService.collectThirdPartyForDeletion(classifiedArtefactMatches);
-        ThirdPartySubscriptionArtefact subscriptionArtefact = new ThirdPartySubscriptionArtefact(
-            TEST, classifiedArtefactMatches);
-        verify(publicationServicesService, never()).sendEmptyArtefact(subscriptionArtefact);
-    }
-
-    @Test
     void testMiServiceLocal() {
         when(subscriptionRepository.getLocalSubsDataForMi()).thenReturn(EXAMPLE_CSV_LOCAL);
         String testString = subscriptionService.getLocalSubscriptionsDataForMiReporting();
@@ -748,33 +289,6 @@ class SubscriptionServiceTest {
             .as("Wrong comma count compared to header row!")
             .allSatisfy(
                 e -> assertThat(e.chars().filter(character -> character == ',').count()).isEqualTo(countLine1));
-    }
-
-    @Test
-    void testDeleteAllByUserId() {
-        String testString = "testId";
-        ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        doNothing().when(subscriptionRepository).deleteAllByUserId(captor.capture());
-        subscriptionService.deleteAllByUserId(testString);
-        assertEquals(testString, captor.getValue(),
-                     "The service layer failed to delete the correct user id subscriptions");
-    }
-
-    @Test
-    void testFindSubscriptionsByLocationIdReturnsExpected() {
-        when(subscriptionRepository.findSubscriptionsByLocationId(LOCATION_ID))
-            .thenReturn(mockSubscriptionList);
-        assertEquals(mockSubscriptionList, subscriptionService.findSubscriptionsByLocationId(LOCATION_ID),
-                     "The returned subscription list does not match the expected list");
-    }
-
-    @Test
-    void testFindSubscriptionsByLocationIdException() {
-        when(subscriptionRepository.findSubscriptionsByLocationId(LOCATION_ID)).thenReturn(List.of());
-        assertThrows(SubscriptionNotFoundException.class, () ->
-                         subscriptionService.findSubscriptionsByLocationId(LOCATION_ID),
-                     "SubscriptionNotFoundException not thrown "
-                         + "when trying to find a subscription that does not exist");
     }
 }
